OAuth1.0a: send unmodified site response to client
OAuth1.0a allows websites to return additional pairs of information to
the clients, besides the token and the token secret. This possibility is
used at least for Twitter and Flickr, who provide the username and the
user ID to the client.
Previously to this commit, the Twitter-specific fields were added to the
session data reply (and this commit doesn't remove them, form
compatibility reasons); however, adding support to any specific site is
not scalable. So, in addition to the previously used fields, we now also
pass all the unmodified response fields to the client application.