OAuth2: authorize client with HTTP Basic auth
The OAuth 2.0 specification says that the preferred way to authorize
clients is via the HTTP Basic Authorization:
http://tools.ietf.org/html/rfc6749#section-2.3
This commit makes that the default way of authorizing clients.

However, despite the fact that the specification states that services
MUST support this authentication, the reality is that at least Google
doesn't. For that reason, we add a flag (ForceClientAuthViaRequestBody,
named after the OAuth2 plugin for gsso) which clients can add in the
session data to force the old behaviour.

Fixes: https://code.google.com/p/accounts-sso/issues/detail?id=232