Delay instantiation of QNetworkAccessManager
signond expects the authentication plugins to be initialized in 5
seconds. However, in some cases instantiating a QNetworkAccessManager
can take even 10 seconds (see https://bugs.launchpad.net/bugs/1377181).

To workaround this kind of situations, we delay instantiating the
QNetworkAccessManager until it's needed in the process() method, which
does not have any time constraints.