OAuth2: implement a fallback mechanism when parsing replies
Some providers, such as Sina Weibo, return a JSON reply while setting
the content type to "text/plain", causing our parsing to fail.
With this change, the content type is taken as a hint only, to select
the best parser; but if the parsing fails, we try anyway to feed the
content to the other parser.