OAuth2: handle errors occurring when refreshing a token
In https://bugs.launchpad.net/bugs/1420934 the user revokes the
application's permission to use the account, while the access token
currently cached in signond is expired and we have a refresh token
available.
When attempting to use the refresh token to get a new access token, we
get a QNetworkReply::error() with a ProtocolInvalidOperationError code;
this causes the plugin to immediately respond to signond with an error
signal. The current code, however, was expecting a normal termination of
the network operation and would handle the error only if it was reported
in the HTTP status code.
The current commit makes the code more robust, handling both the case
when QNetworkReply::error() is emitted before QNetworkReply::finished(),
and also when only one of these signals is emitted (the Qt documentation
says that a finished() signal will *probably* follow).

The changes in BasePlugin ensure that only the first signal between
finished() and error() will be handled, and that in any case the reply
will be checked for errors.